Interval override handling
This change is for specific HID Remapper needs.

It makes it possible to override the polling interval (rate
overclocking).

Also, when not overriding the interval, it restricts the interval
declared by the device in its endpoint descriptor to powers of two,
which is what computers do normally (but TinyUSB doesn't).